css-text-3_dev/html/overflow-wrap-break-word-fit-content-001.htm

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>CSS Text Test: overflow-wrap: break-word with fit-content</title>
<link href="kojiishi@gmail.com" rel="author" title="Koji Ishii">
<link href="http://www.w3.org/TR/css-text-3/#overflow-wrap" rel="help" title="6.2. Overflow Wrapping: the 'word-wrap'/'overflow-wrap' property">
<meta content="dom" name="flags">
<meta content="The 'overflow-wrap: break-word' must not break surrogate pairs" name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.fit {
  display: inline-block;
  word-wrap: break-word;
  font-size: 20px;
  line-height: 1;
}
</style>
</head><body>
<div id="log"></div>
<p classs="instructions">Tests pass if all lines do not wrap.
</p><ol id="container">
  <li><div class="fit">ğœ‹</div>
  </li><li><div class="fit">ğœ‹</div>
</li></ol>
<script>
var tests = document.querySelectorAll(".fit");
for (var i = 0; i < tests.length; i++) {
  var element = tests[i];
  test(function () {
    var rects = element.getClientRects();
    assert_equals(1, rects.length);
    assert_equals(rects[0].height, 20);
  }, (i + 1).toString());
}
</script>

</body></html>
